+++ lib/src/extension/duration_extension.dart
@@ -1,12 +1,26 @@
 import 'package:rpmtw_dart_common_library/src/util/rpmtw_util.dart';
 
 extension DurationExtension on Duration {
-  String format(
-          {required String i18nHours,
-          required String i18nMinutes,
-          required String i18nSeconds}) =>
+  /// See also [RPMTWUtil.formatDuration]
+  String format({
+    required String i18nDay,
+    required String i18nHour,
+    required String i18nMinute,
+    required String i18nSecond,
+    bool forceShowSeconds = false,
+    bool forceShowMinutes = false,
+    bool forceShowHours = false,
+    bool forceShowDays = false,
+    bool abbreviate = false,
+  }) =>
       RPMTWUtil.formatDuration(this,
-          i18nHours: i18nHours,
-          i18nMinutes: i18nMinutes,
-          i18nSeconds: i18nSeconds);
+          i18nDay: i18nDay,
+          i18nHour: i18nHour,
+          i18nMinute: i18nMinute,
+          i18nSecond: i18nSecond,
+          forceShowSeconds: forceShowSeconds,
+          forceShowMinutes: forceShowMinutes,
+          forceShowHours: forceShowHours,
+          forceShowDays: forceShowDays,
+          abbreviate: abbreviate);
 }