+++ lib/src/logger/logger_level.dart
@@ -1,0 +1,11 @@
+/// [LoggerLevel]s to control logging output. Logging can be enabled to include all
+/// levels above certain [LoggerLevel].
+enum LoggerLevel {
+  verbose,
+  debug,
+  info,
+  warning,
+  error,
+  fatal,
+  nothing,
+}