+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.4
+- feat: Logger level can be customized.
+- feat: more option for `formatDuration()` method.
+- feat: added `getUTCTime()` method to RPMTWUtil.
+- build: set minimum dart sdk version to 2.12.0.
+- docs: added docs for all functions.
+- test: added all tests.
 # 0.0.3
 - feat: added `formatDuration()` method to RPMTWUtil
 ## 0.0.2